starter-apps/streamlit-apps/SoulSpeak - AI Buddy/requirements.txt
streamlit==1.46.0
python-dotenv>=1.0.0
agno>=0.1.6
langdetect>=1.0.9
SpeechRecognition>=3.10.0
textwrap3>=0.9.2
tavily-python>=0.3.3
openai>=1.86.0
google-genai>=1.23.0